+++ shuup/customer_group_pricing/models.py
@@ -15,16 +15,23 @@ from shuup.core.utils.context_cache import bump_cache_for_product
 from shuup.utils.properties import MoneyPropped, PriceProperty
 
 
-class CgpPrice(MoneyPropped, models.Model):
+class CgpBase(models.Model):
     product = models.ForeignKey("shuup.Product", related_name="+", on_delete=models.CASCADE, verbose_name=_("product"))
     shop = models.ForeignKey("shuup.Shop", db_index=True, on_delete=models.CASCADE, verbose_name=_("shop"))
     group = models.ForeignKey(
         "shuup.ContactGroup", db_index=True, on_delete=models.CASCADE, verbose_name=_("contact group"))
+
+    class Meta:
+        abstract = True
+        unique_together = (('product', 'shop', 'group'),)
+
+
+class CgpPrice(MoneyPropped, CgpBase):
     price = PriceProperty("price_value", "shop.currency", "shop.prices_include_tax")
     price_value = MoneyValueField(verbose_name=_("price"))
 
-    class Meta:
-        unique_together = (('product', 'shop', 'group'),)
+    class Meta(CgpBase.Meta):
+        abstract = False
         verbose_name = _(u"product price")
         verbose_name_plural = _(u"product prices")
 
@@ -38,6 +45,31 @@ class CgpPrice(MoneyPropped, models.Model):
 
     def save(self, *args, **kwargs):
         super(CgpPrice, self).save(*args, **kwargs)
+
+        # check if there is a shop product before bumping the cache
+        if self.product.shop_products.filter(shop_id=self.shop.id).exists():
+            bump_cache_for_product(self.product, self.shop)
+
+
+class CgpDiscount(MoneyPropped, CgpBase):
+    discount_amount = PriceProperty("discount_amount_value", "shop.currency", "shop.prices_include_tax")
+    discount_amount_value = MoneyValueField(verbose_name=_("discount amount"))
+
+    class Meta(CgpBase.Meta):
+        abstract = False
+        verbose_name = _(u"product discount")
+        verbose_name_plural = _(u"product discounts")
+
+    def __repr__(self):
+        return "<CgpDiscount (p%s,s%s,g%s): discount %s" % (
+            self.product_id,
+            self.shop_id,
+            self.group_id,
+            self.discount_amount
+        )
+
+    def save(self, *args, **kwargs):
+        super(CgpDiscount, self).save(*args, **kwargs)
 
         # check if there is a shop product before bumping the cache
         if self.product.shop_products.filter(shop_id=self.shop.id).exists():