+++ shuup_tests/customer_group_pricing/test_form.py
@@ -10,8 +10,8 @@ from django.conf import settings
 
 from shuup.core.models import ContactGroup, Shop
 from shuup.customer_group_pricing.admin_form_part import \
-    CustomerGroupPricingForm
-from shuup.customer_group_pricing.models import CgpPrice
+    CustomerGroupPricingForm, CustomerGroupDiscountForm
+from shuup.customer_group_pricing.models import CgpDiscount, CgpPrice
 from shuup.testing.factories import (
     create_product, get_default_customer_group, get_default_shop
 )
@@ -29,33 +29,40 @@ def _get_test_product():
     return product
 
 
+@pytest.mark.parametrize("form", [CustomerGroupPricingForm, CustomerGroupDiscountForm])
 @pytest.mark.django_db
-def test_basic_form_sanity():
+def test_basic_form_sanity(form):
     shop = get_default_shop()
     group = get_default_customer_group()
     product = _get_test_product()
-    frm = CustomerGroupPricingForm(product=product, empty_permitted=True)
+    frm = form(product=product, empty_permitted=True)
     assert len(frm.groups) == ContactGroup.objects.count()
     assert len(frm.shops) == Shop.objects.count()
 
     assert "s_%d_g_%d" % (shop.id, group.id) in frm.fields
 
 
+@pytest.mark.parametrize("form", [CustomerGroupPricingForm, CustomerGroupDiscountForm])
 @pytest.mark.django_db
-def test_no_changes_into_form():
+def test_no_changes_into_form(form):
     product = _get_test_product()
     shop = get_default_shop()
-    frm = CustomerGroupPricingForm(product=product, empty_permitted=True)
+    frm = form(product=product, empty_permitted=True)
     # No changes made, right?
     form_data = get_form_data(frm, prepared=True)
-    frm = CustomerGroupPricingForm(product=product, data=form_data, empty_permitted=True)
+    frm = form(product=product, data=form_data, empty_permitted=True)
     frm.full_clean()
     frm.save()
-    assert CgpPrice.objects.get(product=product, shop=shop).price.value == 250
+
+    if form == CustomerGroupPricingForm:
+        assert CgpPrice.objects.get(product=product, shop=shop).price.value == 250
+    else:
+        assert CgpDiscount.objects.filter(product=product, shop=shop).count() == 0
 
 
+@pytest.mark.parametrize("form", [CustomerGroupPricingForm, CustomerGroupDiscountForm])
 @pytest.mark.django_db
-def test_change_shop_price():
+def test_change_shop_price(form):
     product = _get_test_product()
     shop = get_default_shop()
     group = get_default_customer_group()
@@ -63,30 +70,53 @@ def test_change_shop_price():
 
     form_field = "s_%d_g_%d" % (shop.id, group.id)
 
-    frm = CustomerGroupPricingForm(product=product, empty_permitted=True)
+    frm = form(product=product, empty_permitted=True)
     form_data = get_form_data(frm, prepared=True)
-    # Price hike time!
-    form_data[form_field] = "4000"
-    frm = CustomerGroupPricingForm(product=product, data=form_data, empty_permitted=True)
+
+    if form == CustomerGroupPricingForm:
+        # Price hike time!
+        form_data[form_field] = "4000"
+    else:
+        form_data[form_field] = "200"
+
+    frm = form(product=product, data=form_data, empty_permitted=True)
     frm.full_clean()
     frm.save()
-    assert CgpPrice.objects.get(product=product, shop=shop, group=group).price == price(4000)
+
+    if form == CustomerGroupPricingForm:
+        assert CgpPrice.objects.get(product=product, shop=shop, group=group).price == price(4000)
+    else:
+        assert CgpDiscount.objects.get(product=product, shop=shop, group=group).discount_amount == price(200)
 
     # Never mind actually, same price for all shops
     form_data[form_field] = ""
-    frm = CustomerGroupPricingForm(product=product, data=form_data, empty_permitted=True)
+    frm = form(product=product, data=form_data, empty_permitted=True)
     frm.full_clean()
     frm.save()
 
-    assert not CgpPrice.objects.filter(product=product, shop=shop, group=group).exists()
+    if form == CustomerGroupPricingForm:
+        assert not CgpPrice.objects.filter(product=product, shop=shop, group=group).exists()
+    else:
+        assert not CgpDiscount.objects.filter(product=product, shop=shop, group=group).exists()
 
 
+@pytest.mark.parametrize("form", [CustomerGroupPricingForm, CustomerGroupDiscountForm])
 @pytest.mark.django_db
-def test_clear_prices():
+def test_clear_prices(form):
     product = _get_test_product()
     # We can clear the prices out, can't we?
     form_data = {}
-    frm = CustomerGroupPricingForm(product=product, data=form_data, empty_permitted=True)
+
+    if form == CustomerGroupDiscountForm:
+        group = get_default_customer_group()
+        shop = get_default_shop()
+        CgpDiscount.objects.create(product=product, shop=shop, group=group, discount_amount_value=10)
+
+    frm = form(product=product, data=form_data, empty_permitted=True)
     frm.full_clean()
     frm.save()
-    assert not CgpPrice.objects.filter(product=product).exists()
+
+    if form == CustomerGroupPricingForm:
+        assert not CgpPrice.objects.filter(product=product).exists()
+    else:
+        assert not CgpDiscount.objects.filter(product=product, shop=shop).exists()